# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your super  ApplicationAPI",
        "version": "1.0.0"
    },
    "paths": {
...
        "http://80.78.246.59/Stech/public/api/createMessage": {
            "post": {
                "tags": [
                    "Message"
                ],
                "summary": "createMessage",
                "description": "createMessage",
                "operationId": "createMessage",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "phone"
                                ],
                                "properties": {
                                    "receiver_id": {
                                        "type": "string",
                                        "format": "receiver_id",
                                        "example": "1"
                                    },
                                    "message": {
                                        "type": "string",
                                        "format": "message",
                                        "example": "Грузо-перевозки"
                                    },
                                    "file": {
                                        "type": "file",
                                        "format": "file",
                                        "example": "photo.png"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "http://80.78.246.59/Stech/public/api/ChatUser": {
            "get": {
                "tags": [
                    "Message"
                ],
                "summary": "Get list of ChatUser",
                "description": "Returns list of ChatUser",
                "operationId": "ChatUser",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "http://80.78.246.59/Stech/public/api/OnePageMessage": {
            "post": {
                "tags": [
                    "Message"
                ],
                "summary": "OnePageMessage",
                "description": "OnePageMessage",
                "operationId": "OnePageMessage",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "phone"
                                ],
                                "properties": {
                                    "receiver_id": {
                                        "type": "string",
                                        "format": "receiver_id",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "http://80.78.246.59/Stech/public/api/addNewInvite": {
            "post": {
                "tags": [
                    "Invite"
                ],
                "summary": "addNewInvite",
                "description": "addNewInvite",
                "operationId": "addNewInvite",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "phone"
                                ],
                                "properties": {
                                    "receiver_id": {
                                        "type": "string",
                                        "format": "receiver_id",
                                        "example": "1"
                                    },
                                    "tender_id": {
                                        "type": "string",
                                        "format": "tender_id",
                                        "example": "2"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "http://80.78.246.59/Stech/public/api/getRoleId3Invite": {
            "get": {
                "tags": [
                    "Invite"
                ],
                "summary": "Get list of getRoleId3Invite",
                "description": "Returns list of getRoleId3Invite",
                "operationId": "getRoleId3Invite",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "http://80.78.246.59/Stech/public/api/SuccsesInvite": {
            "post": {
                "tags": [
                    "Invite"
                ],
                "summary": "SuccsesInvite",
                "description": "SuccsesInvite",
                "operationId": "SuccsesInvite",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "phone"
                                ],
                                "properties": {
                                    "sender_id": {
                                        "type": "string",
                                        "format": "sender_id",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "http://80.78.246.59/Stech/public/api/deleteInvite": {
            "post": {
                "tags": [
                    "Invite"
                ],
                "summary": "deleteInvite",
                "description": "deleteInvite",
                "operationId": "deleteInvite",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "phone"
                                ],
                                "properties": {
                                    "sender_id": {
                                        "type": "string",
                                        "format": "sender_id",
                                        "example": "1"
                                    },
                                    "invite_id": {
                                        "type": "string",
                                        "format": "invite_id",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
...
    }
}